+++ myokit/tests/test_formats_exporters_run.py
@@ -131,6 +131,9 @@ class ExportTest(unittest.TestCase):
     def test_cuda_kernel_rl_exporter(self):
         self._test(myokit.formats.exporter('cuda-kernel-rl'))
 
+    def test_diffsl_exporter(self):
+        self._test(myokit.formats.exporter('diffsl'))
+
     def test_easyml_exporter(self):
         self._test(myokit.formats.exporter('easyml'))