+++ CHANGELOG.md
@@ -4,13 +4,22 @@ This page lists the main changes made to Myokit in each release.
 
 ## Unreleased
 - Added
+- Changed
+  - [#1102](https://github.com/myokit/myokit/pull/1102) Replaced the HEKA method `c_fast_tau` with a method `c_fast_detailed`.
+- Deprecated
+- Removed
+  - [#1111](https://github.com/myokit/myokit/pull/1111) No longer testing on Python 3.7.
+- Fixed
+  - [#1100](https://github.com/myokit/myokit/pull/1100) Added missing PatchMaster classes to the API and docs.
+  - [#1109](https://github.com/myokit/myokit/pull/1109) Fixed issue with meta data in Patchmaster class for files with alternative amplifier state storage.
+
+## [1.37.1] - 2024-11-05
+- Added
+  - [#1073](https://github.com/myokit/myokit/pull/1073) Added export to [DiffSL](https://github.com/martinjrobins/diffeq-js/).
   - [#1088](https://github.com/myokit/myokit/pull/1088) The DataLogViewer now displays meta data for myokit DataLogs.
   - [#1088](https://github.com/myokit/myokit/pull/1088) The HEKA file format now supports more meta data, including filters.
   - [#1092](https://github.com/myokit/myokit/pull/1092) The method `DataLog.load_csv` can now read CSV files with duplicate column names.
   - [#1095](https://github.com/myokit/myokit/pull/1095) Myokit is now tested on Python 3.12 and 3.13.
-- Changed
-- Deprecated
-- Removed
 - Fixed
   - [#1083](https://github.com/myokit/myokit/pull/1083) Fixed bug when passing initial state `Expression` objects to `myokit.step`.
   - [#1088](https://github.com/myokit/myokit/pull/1088) The HEKA file format meta data now shows the correct value for fast capacitance correction.